fix: rename miniDiff to mini_diff for consistency
